Refactor gdalinfo into separate submodule
